image_processing/requirements.txt
opencv-python
face_recognition
requests
python-dotenv
numpy
flask
